implementing the form changes
