+++ src/app/data/counties .json
@@ -1,0 +1,229 @@
+{
+  "counties": [
+    "Aberdeen City",
+    "Aberdeenshire",
+    "Angus",
+    "Antrim and Newtownabbey",
+    "Ards and North Down",
+    "Argyll and Bute",
+    "Armagh, Banbridge and Craigavon",
+    "Barking and Dagenham",
+    "Barnet",
+    "Barnsley",
+    "Bath and North East Somerset",
+    "Bedford",
+    "Belfast",
+    "Bexley",
+    "Birmingham",
+    "Blackburn with Darwen",
+    "Blackpool",
+    "Blaenau Gwent",
+    "Bolton",
+    "Bournemouth",
+    "Bracknell Forest",
+    "Bradford",
+    "Brent",
+    "Bridgend [Pen-y-bont ar Ogwr]",
+    "Brighton and Hove",
+    "Bristol, City of",
+    "Bromley",
+    "Buckinghamshire",
+    "Bury",
+    "Caerphilly [Caerffili]",
+    "Calderdale",
+    "Cambridgeshire",
+    "Camden",
+    "Cardiff [Caerdydd]",
+    "Carmarthenshire [Sir Gaerfyrddin]",
+    "Causeway Coast and Glens",
+    "Central Bedfordshire",
+    "Ceredigion [Sir Ceredigion]",
+    "Cheshire East",
+    "Cheshire West and Chester",
+    "Clackmannanshire",
+    "Conwy",
+    "Cornwall",
+    "Coventry",
+    "Croydon",
+    "Cumbria",
+    "Darlington",
+    "Denbighshire [Sir Ddinbych]",
+    "Derby",
+    "Derbyshire",
+    "Derry and Strabane",
+    "Devon",
+    "Doncaster",
+    "Dorset",
+    "Dudley",
+    "Dumfries and Galloway",
+    "Dundee City",
+    "Durham County",
+    "Ealing",
+    "East Ayrshire",
+    "East Dunbartonshire",
+    "East Lothian",
+    "East Renfrewshire",
+    "East Riding of Yorkshire",
+    "East Sussex",
+    "Edinburgh, City of",
+    "Eilean Siar",
+    "Enfield",
+    "England",
+    "England and Wales",
+    "Essex",
+    "Falkirk",
+    "Fermanagh and Omagh",
+    "Fife",
+    "Flintshire [Sir y Fflint]",
+    "Gateshead",
+    "Glasgow City",
+    "Gloucestershire",
+    "Great Britain",
+    "Greenwich",
+    "Gwynedd",
+    "Hackney",
+    "Halton",
+    "Hammersmith and Fulham",
+    "Hampshire",
+    "Haringey",
+    "Harrow",
+    "Hartlepool",
+    "Havering",
+    "Herefordshire",
+    "Hertfordshire",
+    "Highland",
+    "Hillingdon",
+    "Hounslow",
+    "Inverclyde",
+    "Isle of Anglesey [Sir Ynys Môn]",
+    "Isle of Wight",
+    "Isles of Scilly",
+    "Islington",
+    "Kensington and Chelsea",
+    "Kent",
+    "Kingston upon Hull",
+    "Kingston upon Thames",
+    "Kirklees",
+    "Knowsley",
+    "Lambeth",
+    "Lancashire",
+    "Leeds",
+    "Leicester",
+    "Leicestershire",
+    "Lewisham",
+    "Lincolnshire",
+    "Lisburn and Castlereagh",
+    "Liverpool",
+    "London, City of",
+    "Luton",
+    "Manchester",
+    "Medway",
+    "Merthyr Tydfil [Merthyr Tudful]",
+    "Merton",
+    "Mid and East Antrim",
+    "Mid Ulster",
+    "Middlesbrough",
+    "Midlothian",
+    "Milton Keynes",
+    "Monmouthshire [Sir Fynwy]",
+    "Moray",
+    "Neath Port Talbot [Castell-nedd Port Talbot]",
+    "Newcastle upon Tyne",
+    "Newham",
+    "Newport [Casnewydd]",
+    "Newry, Mourne and Down",
+    "Norfolk",
+    "North Ayrshire",
+    "North East Lincolnshire",
+    "North Lanarkshire",
+    "North Lincolnshire",
+    "North Somerset",
+    "North Tyneside",
+    "North Yorkshire",
+    "Northamptonshire",
+    "Northern Ireland",
+    "Northumberland",
+    "Nottingham",
+    "Nottinghamshire",
+    "Oldham",
+    "Orkney Islands",
+    "Oxfordshire",
+    "Pembrokeshire [Sir Benfro]",
+    "Perth and Kinross",
+    "Peterborough",
+    "Plymouth",
+    "Poole",
+    "Portsmouth",
+    "Powys",
+    "Reading",
+    "Redbridge",
+    "Redcar and Cleveland",
+    "Renfrewshire",
+    "Rhondda, Cynon, Taff [Rhondda, Cynon,Taf]",
+    "Richmond upon Thames",
+    "Rochdale",
+    "Rotherham",
+    "Rutland",
+    "Salford",
+    "Sandwell",
+    "Scotland",
+    "Scottish Borders, The",
+    "Sefton",
+    "Sheffield",
+    "Shetland Islands",
+    "Shropshire",
+    "Slough",
+    "Solihull",
+    "Somerset",
+    "South Ayrshire",
+    "South Gloucestershire",
+    "South Lanarkshire",
+    "South Tyneside",
+    "Southampton",
+    "Southend-on-Sea",
+    "Southwark",
+    "St. Helens",
+    "Staffordshire",
+    "Stirling",
+    "Stockport",
+    "Stockton-on-Tees",
+    "Stoke-on-Trent",
+    "Suffolk",
+    "Sunderland",
+    "Surrey",
+    "Sutton",
+    "Swansea [Abertawe]",
+    "Swindon",
+    "Tameside",
+    "Telford and Wrekin",
+    "Thurrock",
+    "Torbay",
+    "Torfaen [Tor-faen]",
+    "Tower Hamlets",
+    "Trafford",
+    "United Kingdom",
+    "Vale of Glamorgan, The [Bro Morgannwg]",
+    "Wakefield",
+    "Wales [Cymru]",
+    "Walsall",
+    "Waltham Forest",
+    "Wandsworth",
+    "Warrington",
+    "Warwickshire",
+    "West Berkshire",
+    "West Dunbartonshire",
+    "West Lothian",
+    "West Sussex",
+    "Westminster",
+    "Wigan",
+    "Wiltshire",
+    "Windsor and Maidenhead",
+    "Wirral",
+    "Wokingham",
+    "Wolverhampton",
+    "Worcestershire",
+    "Wrexham [Wrecsam]",
+    "York"
+  ]
+}
+
